*** empty log message ***
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@511 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/peerlist-template_impl.h
@@ -198,7 +198,7 @@ void ST_CLASS(peerListGetDescription)(struct ST_CLASS(PeerList)* peerList,
 {
    snprintf(buffer, bufferSize, "PeerList of $%08x (%u peers)",
             peerList->OwnIdentifier,
-            ST_CLASS(peerListGetPeerListNodes)(peerList));
+            (unsigned int)ST_CLASS(peerListGetPeerListNodes)(peerList));
 }
